Some fixes but not all
